Creación del endpoint para cuenta, se agrega logica de negocio para espacios y cultivos

+++ crops.js
@@ -104,6 +104,9 @@ router.put('/update/:id', async (req, res) => {
 
     try {
         const cropRef = db.collection('crops').doc(id);
+        const cropDoc = await cropRef.get();
+        const oldName = cropDoc.data().name; // Nombre antiguo del cultivo
+
         await cropRef.update({
             name,
             crop,
@@ -116,6 +119,21 @@ router.put('/update/:id', async (req, res) => {
             type,
             space
         })
+
+        const deviceRef = db.collection('devices');
+        const deviceQuery = await deviceRef.where('crop', '==', oldName).get();
+
+        if (!deviceQuery.empty) {
+            const batch = db.batch(); // Usar un batch para múltiples actualizaciones
+
+            deviceQuery.forEach((doc) => {
+                const deviceDocRef = db.collection('devices').doc(doc.id);
+                batch.update(deviceDocRef, { crop: name }); // Actualizar el campo "crop"
+            });
+
+            await batch.commit(); // Ejecutar todas las actualizaciones en lote
+        }
+
         return res.status(201).send({ msg: 'Cultivo actualizado correctamente' });
     } catch (error) {
         return res.status(500).send({ msg: 'Error al actualizar el cultivo' });
@@ -125,11 +143,39 @@ router.put('/update/:id', async (req, res) => {
 router.delete('/delete/:id', async (req, res) => {
     try {
         const { id } = req.params;
+
+        // Obtener el cultivo antes de eliminarlo para obtener su nombre
+        const cropDoc = await db.collection('crops').doc(id).get();
+        if (!cropDoc.exists) {
+            return res.status(404).send({ msg: "El cultivo no existe" });
+        }
+
+        const cropName = cropDoc.data().name; // Nombre del cultivo
+
+        // Eliminar el cultivo
         await db.collection('crops').doc(id).delete();
+
+        // Buscar todos los dispositivos asociados al cultivo
+        const deviceRef = db.collection('devices');
+        const deviceQuery = await deviceRef.where('crop', '==', cropName).get();
+
+        // Si hay dispositivos asociados, eliminarlos
+        if (!deviceQuery.empty) {
+            const batch = db.batch(); // Usar un batch para múltiples eliminaciones
+
+            deviceQuery.forEach((doc) => {
+                const deviceDocRef = db.collection('devices').doc(doc.id);
+                batch.delete(deviceDocRef); // Eliminar el dispositivo
+            });
+
+            await batch.commit(); // Ejecutar todas las eliminaciones en lote
+        }
+
         return res.status(200).send({ msg: "Cultivo eliminado correctamente" });
     } catch (error) {
-        return res.status(500).send({ msg: 'Error al eliminar el cultivo' });
+        console.error("ERROR DELETE /delete/:id", error);
+        return res.status(500).send({ msg: "Error al eliminar el cultivo" });
     }
-})
+});
 
 module.exports = router;
